Enable automatic migration of /etc contents (copy)
Copy the contents of the existing /etc overlay file system into the
snapshot; this will enable RPM scripts which are trying to modify
existing configuration files to be able to do so now (up until now
they just didn't see the files in the overlay).
On first boot after a new snapshot generation delete the contents of
the overlay to avoid inconsistencies with older snapshots. This will
loose all changes to /etc since the last snapshot though, and old
snapshots will get the changes accumulated since the last snapshot.