Introduce snapshot specific /etc overlays
This commit resumes the idea of adding a new overlay directory for
each snapshot.

To do so this approach is using dracut's feature to read /etc/fstab.sys
from the root file system. The previous overlay directories will just
be added to the /etc line.

Together with followup patches this is supposed to create a hybrid of
the two POCs for improved /etc handling. By combining those two
approaches it is possible to roll back to a previous snapshot even
when /etc is in an inconsistent state, as for previous snapshots the
latest /etc overlay(s) are not applied - administrators repeatedly
broke their installations by creating invalid /etc configurations.
Also a lot of the special handling caused by sharing one /etc for all
snapshots can be removed. This will also significantly decrease the
complexity of the dracut service, as the overlay cleanup does not have
to be done by hand any more - it will only print a warning on
conflicting changes now. By combining it with the copy approach the
infinite growing of the overlays is stopped and the complexity of the
necessary cleanup algorithm is decreased significantly.

Note that this commit alone will create a large stack of overlays over
time.