Avoid warning when copying network config
On existing systems /var/run may (accidentally) be part of the snapshot. If
so it should be a symlink pointing to /run, so copying the configuration file
will be refused as the source and target are the same file.
To avoid that just delete the directory if /var is a separate subvolume - the
directory will be deleted at the end of the transactional-update run anyway.

Old systems with the non-unified /var subvolume layout are not affected.