Move update checking code
Previously a zypper --dry-run was executed before setting up the btrfs
snapshot. If no update was pending the application would simply exit again.

That behaviour to two problems:
* Non-interactive updates didn't find the packages if a new repository was
  added and a new key would have to be imported - even when the option
  ZYPPER_AUTO_IMPORT_KEYS was set
* Only non-interactive commands were checked - but if a noop was triggered
  in interactive mode (e.g. installing an already existing package) that
  wouldn't be detected

To resolve this the update detection code has been moved and is executed
after snapshot creation now. If nothing was done the snapshot is simply
deleted again (like in the error case, but with a success return value).

This also made is possible to get rid of the complex XML error message
parsing, as the dry-run is only used for checking the number of package
changes. This makes it possible to get rid both of the Perl and xsltproc
dependencies.