Rework messages on failed umounts
If the update succeeded, but the unmount of one of the bind mounts failed,
it will be attempted a second time to unmount the mount point. The error
message however will appear on the first attempt, even if the second
attempt was successful.
In combination with [boo#1168389] this behaviour is unfortunate, because
this may lead to the assumption that the complete update failed, so this
commit reworks the messages to only indicate a hard error if the second
attempt also didn't succeed.