Reimplement execution of commands
Rename the "Transaction" methods into "init" and "finalize" to better describe
the lifecycle of the transaction. If the destructor should be called before
finalize has been called the transactional will be rolled back.
The output of the command to be executed will accept input and print output
directly now. This allows one to e.g. call a `bash` instance in a transactional
update environment.