Reduce number of overlays
Instead of mounting all overlays from all previous snapshots sycnronize the
previous snapshot's /etc state (instead of the oldest one) and only add the
current snapshot's overlay to the new fstab.

This will greatly reduce the number of overlays and will also avoid confusion,
as changes to any of the older overlays (e.g. when temporarily booting an old
snapshot) were also applied to the current system. Now only changes made
to the system the current snapshot was branched from will be applied.