Check mount option length to not exceed maximum
While overlayfs doesn't have a maximum number of layers, the `mount` syscall
does have a length limit for the option string, having to fit into one
PAGE_SIZE.
The limit may be exceeded if a lot of new snapshots (>100) are created using
the --continue option without rebooting the system. Abort and ask the user
to reboot before continuing in that case.