Remove ro protection of /var/lib/ca-certificates
The directory is part of a package; on package upgrades SELinux will try to
change the directory's attributes, which will fail if there are no write
permissions.
Packages are not supposed to operate on this directory directly (the directory
is maintained by `update-ca-certificates`), so accept the risk of modifying
the running system.