Reorganize functions to get mnt_fs handle
The Mount class can be used in two different ways: Once for mount points which
are listed in fstab or mtab, and once for temporary mounts like bind mounts.
The previous implmeentation assumed that the mnt_fs struct could be reused
between these two types, but in practice that didn't work. So reflect that
separate the code paths for mnt_fs retrieval and rename the methods
accordingly.