Use getopt for command line parsing
This changes the syntax for optional parameters (in this case the --continue
option) slightly, which has to be given with -c<num> or --continue=<num> now.

Unfortunately getopt potentially also modifies the order of the command line
arguments (not in this case tough), so argv can't be const any more.