Fix SELinux /etc synchronization
The previous commits tried to solve various aspects of /etc synchronization
with SELinux, but they were ignoring the complete use case.

transactional-update will synchronize the /etc state of the previous snapshot
into the next snapshot. Usually - on a SELinux system - the SELinux xattrs have
to be synchronized into the new snapshot of course. Unfortunately this will
fail with unlabelled files (usually when synchronizing an pre-SELinux /etc
state when SELinux is active), as rsync will try to remove the attributes,
which the policy will prevent.

The solution is simple: If the snapshot overlay root directory doesn't have the
SELinux xattrs set (i.e. SELinux was not active at the time of the creation of
the snapshot), then don't sync the SELinux attributes. Otherwise do.
This is possible due to the fact that `transactional-update setup-selinux` will
relabel all files in /etc, so the overlay will contain a complete labelled copy
of /etc; due to that it doesn't matter whether the /etc files of the root file
system are labelled or not.