Keep an active transaction and call zypper with -R
Calling zypper via tukit has the disadvantage that relative paths to RPM
files are not supported any more. To restore that functionality zypper
will be called with the --root option again, however for this an active
tukit session has to be running for having all the required mounts in
place. This is achieved by opening a dummy session with a long sleep
which is killed when the transaction is finished.
This is obviously just a workaround and will need an official
interface in the future.