Overlay: Don't try sync if parent does not exist
If the parent of a snapshot does not exist any more, then don't try to
synchronize the snapshot overlay into the new root.

In theory this could lead to a problem with accumulating layers in case
there is always only one snapshot available (e.g. if the previous
snapshot should always get deleted immediately after a successful boot
into the next snapshot) or in the very theoretical case that someone
always resumes the oldest snapshot.