Use pthread for processing command
The single threaded implementation doesn't work for when the output gets
longer than the buffer of the pipe: The called application will just be
blocked then until someone empties the pipe again.
fork() is not really useful here, as it also clones the transaction object,
so use pthreads for threading.