Use `tukit call` for up, dup and patch
In case zypper has updated itself (exit code 103), then it should still be
possible to install all updates without havin to reboot the system first
[bsc#1185226]. For the `pkg` commands `callext` has to be used to be able
to access files from the currently running file system, but for the up,
dup or patch commands zypper can be called from within the chroot
environment so that zypper will continue the installation of the remaining
packages. Change the call accordingly.