t-u: Purge kernels as part of package operations
By default on transactional systems multiversioning is turned of, but it
can still be turned on by the user. In that case the kernels would never
be purged again, as the corresponding systemd service to do so during
bootup would fail because of the read-only file system.

Fix this by purging the kernels during the update already - in contrast to
non-transactional systems the current kernel doesn't have to be kept as
the snapshot with the update is not active yet.

This also fixes live patching on SLE which requires multiversioning due
to the current packaging. [bsc#1189728]