Fix error message with --drop-if-no-change
When generating the list of shadowed files, awk wouldn't find the file in
the snapshot any more if the snapshot was already deleted. This would
result in an error message, also breaking Salt.

Skip all the irrelevant parts, also avoiding a bogus message to reboot the
system, which is not necessary in this case.