Suppress any unexpected console output with quiet
The --quiet option is often used to just get the output of commands. If
transactional-update itself is also adding to this output, then parsing
that output may fail, e.g. like it just happened with the Salt
integration where a non-capture error message in the transactional-update
script managed to break the JSON output.
To solve this suppress any output by redirecting stdout and stderr in
--quiet mode, and use the original fd copies only when using the
specific command.

Also fixes [gh#openSUSE/transactional-update#73].