tukit: Ignore all mounted dirs for inotify
Just skipping the mount points while registering the inotify handlers is
not enough: Some applications will still write to /run or /var (which are
not mounted within the transactional-update environment, so this would
trigger the notifier.

Reimplement the old behaviour to ignore all directories which are mounted
on the currently running system to avoid this. Additionally mount well
known temporary directories as tmpfs to avoid having to clean up later.