Rework snapshot bind mount mechanism
Previously to set up the update environment everything was mounted into
the /.snapshots directory of the corresponding and then later bind mounted
to a temporary directory for GRUB (which needs a dedicated mount entry to
operate).

This doesn't seem necessary: Just mount the snapshot once into the
temporary directory directly. It has the downside that the snapshot can't
be written to by external applications while it is open, but should never
be done anyway - use the `transactional-update run` or the `tukit call /
callext` commands for that.
Instead there are two upsides: Once it doesn't clutter the mount output
with dozens of duplicates, and second it makes it possible to use
MS_UNBINDABLE for the temporary mount point, preventing unintended
and often destructive or blocking recursive mounts into every new
snapshot; this typically happens in case of errors (kill -9 or segfaults)
or when opening multiple transactions in parallel - these recursive mounts
are usually difficult to get out of again.

An alternative would have been to set up a private mount namespace,
making only the temporary directory visible to the outside. That wouldn't
clutter the mount output any more at all, but it would be more difficult
to debug, as one would have to switch into the transaction's namespace to
still see the actual mounts.