Fix rsyncing /etc into the running system (bsc#1192242)
If a transaction with the "--drop-if-no-change" switch did not perform
changes outside of /, the changes to /etc are rsynced into the running
system (if applicable). However, it used the /etc of the snapshot as
source instead of the open transaction, so without any overlays.

Ultimately, this effectively rolled back any recent changes done to the
system's /etc, producing the referenced bug.

It worked before 832e04da7b ("Rework snapshot bind mount mechanism")
because until then the mount points inside the transaction were also
visible in the snapshot source directories.

Co-authored-by: Alberto Planas <aplanas@suse.de>