Use separate mount namespace for chroot
Prevents mounts from within the chroot environment to influence the tukit
organized mounts. This could lead to the deletion of user data if a mount
overwrote mounts managed by tukit itself (e.g. when using "mount -a" in a
transactional-update shell) and the directories didn't exist in the snapshot
before (so tukit wanted to clean them up again).