create_dirs_from_rpmdb: set SELinux file context of missing directories
During a t-u package installation, the SELinux RPM plugin handles the
contexts of created elements. The /var directory could be located on a
separate partition, which is not mounted in the tu-env. Therefore the
file contexts are lost in the snapshot and need to be set after boot by
the create_dirs_from_rpmdb service.
Fixes openSUSE#84