Remove bogus message on reboot
For any reboot a bogus message that the reboot wasn't successful was triggered
since the reboot part was moved into tukit.
The message will just be removed for now, as it would be difficult to detect
whether the command triggering the reboot returned successfully. There never
was any error handling implemented, so also in the past that message could
never have been triggered...

[gh#openSUSE/transactional-update#92]