Reintroduce separate bind mount in /tmp
Support for ULP (Userspace Live Patching) means that the original process
(using libpulp as a PRELOAD) has to read the patch file from outside the
transactional-update environment. It will thus have to access the patch
in the snapshot directory, but it also means that an open file handle to
that file will persist as long as that process is running.
Unfortunately a few applications (GRUB, findmnt) require that the root
directory of the chroot is a dedicated mount point; it was only recently
changed that snapshot directly is bind mounted on itself for that.
Due to libpulp keeping a file handler open this doesn't work any more, as
the bind mount couldn't be unmounted due to being busy any more. Because
of that partially revert 86fd429f15e4ede9aab1bfa143cc532ad83d24b5 and
use a dedicated mount point in /tmp again.

[jsc#PED-1078]