Optimize umounts
For mount points it's not really necessary to umount each submount
separately, as everything will be unmounted recursively anyway - and in
the update environment all the mounts are submounts of the new snapshot,
so it's enough to just unmount the root file system mount.

Due to that a mount point won't be unmounted by default any more, instead
the root mount has to be flagged using the optional constructor parameter
"umount" now.