tukit: Merge /etc overlay into parent with -d -c
If the new snapshot is a direct descendant of the currently running
system, then /etc changes can be merged back into the currently running
system directly (if there are no changes to the root fs). However if the
parent is some other snapshot, then the /etc changes have to be merged
into the corresponding overlay, which wasn't handled so far.