replace tu-rebuild-kdump-initrd with mkdumprd
Starting with kdump version 1.9.0, the generation of the
kdump initrd is simplified. As a result, kdump find-kernel is no longer
available and thus tu-rebuild-kdump-initrd fails.

To regenerate the initrd, a simple call to mkdumprd is sufficient.

Remove the sbin/tu-rebuild-kdump-initrd script and replace it with
a call to mkdumprd in sbin/transactional-update.