Add support for full syntax sub commands
Some sub commands like `pkg` also have `package`. And since the `up` sub
command is based on `zypper up`, which is an alias for `update`, I
thought of adding both `up` and `update` to this command.

I left the docs as is, since it's just to make sure people don't run
into this mistake or expect this behavior as they can for other sub
commands and in `zypper`. I also did the same for `dup` with the same
reasoning.