tukit: enable kexec's syscall detection feature
On system with lockdown enabled, the kexec_load syscall can't be used
since it doesn't support signature verification, and one need to use the
file based syscall for kexec operation.

Luckily the kexec has a -a/--kexec-syscall-auto option that handles this
situation since mid-2018 and becomes the default recently with
horms/kexec-tools@29fe5067ed. To ensure that the auto detection is
always used even with older kexec-tools, let's just alway specify the -a
argument to kexec.