Always set a cleanup algorithm for snapshots
Previously the cleanup algorithms was only set on a cleanup run, marking
snapshots of previous transactional-update runs. However with the
introduction of the tukit command line application, the D-Bus interface
and the libtukit library these snapshots would never have been added to
the list of previous transactional-update runs. As a consequence these
snapshots will never be cleaned up automatically.

To avoid this always set the cleanup algorithm in the library code
already. The existing cleanup mechanics won't be changed for now, as
those will still be used for rollback snapshot selection.