Determine D-Bus availabilty via system socket
Previously snapper calls would just be retried without D-Bus if they
failed. On non-D-Bus related errors this was always duplicating the log
messages.

Instead just check whether the D-Bus system socket file is there - if so
this is a good enough indicator that the D-Bus service is actually
available.

An alternate implementation could be to actively call Ping() on the
org.freedesktop.DBus.Peer interface, but that would be overkill here.