Fix plugin system to support Bash special chars
The command to be executed was added as parameters for the plugins to parse.
This however caused problems when special Bash characters were included
such as pipes or the dollar sign, causing the plugins to fail.

Fixed by properly encapsulating and escaping the individual command line
options with 'es now.