Support sdbootutil in transactional-update
If sdbootutil is used, some actions like initrd generation or bootloader
installation should happen outside the transaction.

Some other actions, like the rewrite of the GRUB configuration file are
noop.

Also the call to mkdumprd should happen outside the transaction, but due
to some pbl issues this is a noop for now (with a warning message)

Signed-off-by: Alberto Planas <aplanas@suse.com>