Preserve current working directory if possible
When entering the transactional-update environment (especially when using
the `shell` command) the CWD was previously always changed to /.
This was good for predictability, but it was requested to preserve the
current working directory if possible.
To make it more clear where one lands in the current directory is now part
of the shell prompt.

Implements [bsc#1227452].