Remove overlay support for /etc
Instead of creating a transparent overlay for each snapshot, a subvolume in
the snapshot will be created to store /etc. This management part is handled
in a separate snapper plugin, transactional-update itself doesn't have to
know about any specifics any more.

Migrating away from the quite complex overlay handling solves problems such
as [boo#1140472] and [gh#openSUSE/transactional-update#117].
The downside is that detection of changes between snapshot creation and
reboot have to be handled by custom code now. Due to the significantly
reduced complexity, especially when trying to reset or modify a given
/etc state, this is definitely worth it though.