Implement /etc file syncing (rsync POC)
For reference this was a POC based on rsync. The idea was the same as in
the later implementation: Store one reference directory and compare the
contents of both the new snapshot (to be able to exclude those files
from syncing) and the old snapshot (to determine which files to copy) to
it.
However preparing the exclude file lists proved to be extremely complex
already to due rsync's complex format.

This POC only compares files in one directory only - actual directory
handling wasn't implemented any more.