Implement /etc file syncing (Bash POC)
The second attempt was to implement the logic done by rsync ourselves.
This Bash POC tries to outline how this could be done and is actually
feature complete. The only disadvantage is that it is very slow, probably
due to the huge amount of stat program calls (~1 minute to process the
diff between two snapshots).

The logic of this code is the base for the final C++ implementation though.