Adopt tests to obs environment to avoid root perms
The following restrictions exist:
* fakeroot is not available in all releases
* rootlesskit doesn't have the required permissions

The only test requiring root access so far was the one to change to
group ownership of a file. However on obs the abuild user is member of
the "trusted" group, so let's use that one for now. With this change all
tests should run without any root permissions required.

If something should change again it would be an option to only run the
tests not requiring root access during build.