Separate requested / actual reboot method
In the past transactional-update was only printing the configured reboot
method and the minimally required reboot type, however depending on how
it was configured the actual reboot type may have used a different type.

To make it easier to see which method was actually triggered the
minimally required reboot level and the actually triggered command are
printed separately now.

This should make it easier to implement [poo#163352].