Move sdbootutil functionality to general handling
In the initial commit all the BLS / sdbootutil functionality was put
into one single block at the beginning. This however breaks combining
commands and makes processing of complex commands (e.g. setup-fips)
impossible because it will trigger some of the commands on demand.

All the processing is moved to the corresponding places now, i.e. the
BLS specific code is living right next to the non-BLS specific code
(which also makes it more obvious what's happening).

One difference is that for all commands (also for those which wouldn't
require it such as the creation of a new initrd, which is living outside
the snapshot on BLS systems) a new snapshot is created now. Introducing
a somewhat complex mechanism to detect whether only commands not
modifying the root file system just doesn't seem worth it - especially
because these commands can just be executed outside of a transactional
shell if desired.

The only actual change apart from the new places is that the sdbootutil
commands are called with a specific snapshot number now.

This is step one to get [boo#1246013] working; step three and four
require changes to sdbootutil itself and dracut-fips.