Close snapshot on --keep
Previously a failed snapshot would just be kept open, i.e. the read-only
status would not be reset and temporary files would not be cleaned up.
If a user would then set this snapshot as the default snapshot, then t-u
would falsely assume this is a read-write system.
To eliminate this potential cause of errors the snapshot will be closed
(i.e. set read-only on ro systems) now, but not set as the default
snapshot, and the transactional-update-in-progress=yes userdata will
also not be deleted to mark this snapshot as an irregularly created one.